Added function that extracts well ID and associated genes from GPL file
generate_cloneID_to_orf_table reads in/parses a GPL family file and extracts the table that links the well ids to the genes in the sample.

This table is then saved in input_files for easy use later.

clone_to_orf.csv is the generated file - it has two columns: clone ID and ORF