More work on creating the simple edge S matrix
Flushed out the majority of the code for the creation of the simple edge S matrix.

There is still some issue with the compounds that are stored in the reaction not aligning with what is available online. (This code still prints and exists in the create_simple_S_matrix() function)